+++ src/Foreign/JavaScript.hs
@@ -14,8 +14,9 @@ module Foreign.JavaScript (
 
     -- * Server
     serve, defaultConfig, Config(
-          jsPort, jsAddr, jsWindowReloadOnDisconnect
-        , jsCustomHTML, jsStatic, jsLog),
+          jsPort, jsAddr
+        , jsCustomHTML, jsStatic, jsLog
+        , jsWindowReloadOnDisconnect, jsCallBufferMode),
     Server, MimeType, URI, loadFile, loadDirectory,
     Window, getServer, root,
 
@@ -46,6 +47,7 @@ serve
     -> (Window -> IO ())    -- ^ Initialization whenever a client connects.
     -> IO ()
 serve config init = httpComm config $ eventLoop $ \w -> do
+    setCallBufferMode w (jsCallBufferMode config)
     runFunction w $
         ffi "connection.setReloadOnDisconnect(%1)" $ jsWindowReloadOnDisconnect config
     flushCallBuffer w   -- make sure that all `runEval` commands are executed